Azure_scripts/Test_azure_training/Depth_inpainting/requirements.txt
filelock==3.6.0
imageio==2.9.0
kornia==0.6.1
log_utils==0.4.1
loss_landscapes==3.0.6
matplotlib==3.5.0
net_utils==1.3
networks==0.3.7
numpy==1.18.5
opencv_python_headless==4.1.2.30
Pillow==9.1.0
pynvml==11.4.1
pyrsistent==0.18.0
pytorch_msssim==0.2.1
ray==1.12.0
scikit_learn==1.0.2
torch==1.5.1
torchvision==0.6.1
tqdm==4.32.2
